pyproject.toml
[tool.poetry]
name = "pyqtlearn"
version = "0.1.0"
description = ""
authors = ["Hanserwei <2628273921@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyqt6-tools = "^6.4.2.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
